py-service-tools/organize-photos/requirements.txt
Pillow
# ffmpeg/ffprobe é necessário no sistema (não via pip)
